--- app/features/movie/queries/index.ts
@@ -1,2 +1,0 @@
-export * from "./actions";
-export * from "./use-get-movies-by-query";
