+++ app/features/movie/sections/detail-related-section.tsx
@@ -1,4 +1,4 @@
-import { getMoviesByGenre } from "../queries";
+import { getMoviesByGenre } from "../apis";
 import { MainCarousel } from "../components";
 
 interface DetailRelatedSectionProps {
