+++ app/features/search/components/search-input.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { useGetMoviesByQuery } from "@/app/features/movie/queries";
+import { useGetMoviesByQuery } from "@/app/features/movie/apis";
 import { useDebounce, useObserver } from "@/app/shared/hooks";
 
 import { cn } from "@/app/shared/utils";
