+++ app/shared/components/like/like-button.tsx
@@ -1,7 +1,7 @@
 import { cn } from "@/app/shared/utils";
 import { Heart } from "lucide-react";
-import { toggleLikeMovie } from "@/app/features/movie/queries";
-import { toggleLikeReview } from "@/app/features/review/queries";
+import { toggleLikeMovie } from "@/app/features/movie/apis";
+import { toggleLikeReview } from "@/app/features/review/apis";
 
 interface LikeButtonProps {
   id: string;
